+++ setup_test.go
@@ -77,3 +77,18 @@ var messages []string
 func getRandomMessage() string {
 	return messages[rand.Int()%len(messages)]
 }
+
+func addFullMacros(f xylog.Formatter) xylog.Formatter {
+	return f.AddMacro("asctime", "asctime").
+		AddMacro("created", "created").
+		AddMacro("filename", "filename").
+		AddMacro("funcname", "funcname").
+		AddMacro("levelname", "levelname").
+		AddMacro("levelno", "levelno").
+		AddMacro("lineno", "lineno").
+		AddMacro("module", "module").
+		AddMacro("msecs", "msecs").
+		AddMacro("pathname", "pathname").
+		AddMacro("process", "process").
+		AddMacro("relativeCreated", "relativeCreated")
+}